[WIP] Bus refactor: Fixed bugs, Working on operation transparent ALU

+++ src/TinyProc/Memory/RawMemory.cs
@@ -6,7 +6,7 @@ using TinyProc.Processor;
 public class RawMemory
 {
     public readonly uint _words;
-    public uint TotalSizeBits { get { return Register.SYSTEM_WORD_SIZE * _words; } }
+    public uint TotalSizeBits { get { return (uint)Register.SYSTEM_WORD_SIZE * _words; } }
     // A 2D bool array simulating RAM structure
     private readonly bool[,] _data;
